+++ app/src/main/res-menu/gallery/layout/fragment_gallery.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+        xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
         <import type="android.view.View"/>
@@ -17,7 +17,7 @@
             android:id="@+id/swipeRefreshGallery"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:showView="@{safeUnbox(gallery.isLoading)}">
+            app:showView="@{safeUnbox(gallery.isLoading)}">
 
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/recyclerViewGallery"
@@ -31,7 +31,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
-            tools:hideView="@{safeUnbox(gallery.isLoading)}"/>
+            app:hideView="@{safeUnbox(gallery.isLoading)}"/>
 
         <include
             android:id="@+id/layoutGalleryError"
