+++ app/src/main/res-menu/listplace/layout/fragment_listplace.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
@@ -17,7 +18,7 @@
             android:id="@+id/swipeRefreshListPlace"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:showView="@{safeUnbox(listPlace.isLoading)}"
+            app:showView="@{safeUnbox(listPlace.isLoading)}"
             android:visibility="@{safeUnbox(listPlace.isError()) ? View.INVISIBLE : View.VISIBLE }">
 
             <androidx.recyclerview.widget.RecyclerView
@@ -34,7 +35,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
-            tools:hideView="@{safeUnbox(listPlace.isLoading)}"/>
+            app:hideView="@{safeUnbox(listPlace.isLoading)}"/>
 
         <include
             android:id="@+id/layoutError"
